Added constraint to user :id so can have dots.
By default the :id segment of the restful URI don't allow dots in the ids.
This is because it use the dot to identify the format. Anyways, you can work
around this by setting a constraint on the :id segment of the URI as I did in
this commit. In this case it will allow our users to have usernames that
include the dot character.