Moved email config out to yaml file.
I did this so that the configuration file was completely separate from the
actual Ruby code that is used to initialize the smtp settings in ActionMailer.
This is needed so that I can easily do the next step which is to move the mail
config into the puppet management repository.