Fixed ssh key vulnerability.
There was an issue where if a specially crafted ssh public key was submitted
to the system it would allow the user to ssh into tunnel@tmux.me with a full
shell account. Specifically the scenario was if you public key with a partial
key followed by newline or carriage return and an full valid key would do
this. The reason this happened was that it was applying the authorized_keys
command restrictions to the partial key which is invalid and therefore script
and because the newline/carriage return the valid key that follows was being
written to the authorized_keys file without the command restriction.